Enable light/dark in locations
Enable light/dark in locations. Tweak "look" to work with examine. If a
location is dark you will need a light source to see anything. Enabled
light source artefacts and the ability to switch them on or off. At the
moment powered items don't lose power. Will need to sort that later