#504 using Jest + Copilot to migrate tests - testing tools library
OMG, CoPilot is amazing - I pointed it at my tools library, asked it to write tests and then used those test rrestults to catch some unhandled issues around pluralisation grammar.

It's not perfect - there weere some case where what copilot expected as a test result was not what should happen - so as everyone says, you still need a brain - you ahve to treat it as a code review as it can generate a lot  of tests and code *very* quickly.