#524 - re-enable showing location images
1 line change to fix.
Added "setImageName" function to location, artefact, and creature as well so that images can potentially be updated during gameplay. (likely as a result of missions and events,)

a lot of extra work at the moment to add images for damaged and broken vs fixed items  - but could explore that in future.