#527 - add basic JSON file validation
wanted to do this for years.
now JSON.parse no longer keels over I can!

Small validation on startup that simply processes each JSON file in the data folder and ensures it PARSEs successfully.

If this fails,  then we report which files have failed as an error in the log.