#498 - add location synonym support and useful map functions
we can now find a location (or set of locations) by synonym and vice versa (synonyms for a given location), find the closest matching location, and the distance of a location.

For location synonyms We add them as a collection against location names in the map for slightly quicker handling.