#500 remove redundant locations
during development, there was a plan to allow driving around. The game is big enough as it is.

I'll keep the capability in the engine (but it will almost certainly need review and development) but for the standard game map, deleting the roads.

Fixed up tests and made tweaks for other test reliability (including making mapbuilder use file-only mode.)